pyproject.toml
[project]
name = "rerai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.0",
    "cryptography>=44.0.0",
    "deepagents>=0.4.12",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "playwright>=1.58.0",
    "pypdf>=5.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.2.3",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-timeout>=2.4.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "live: marks tests that hit real external APIs (can be slow or flaky)",
]
testpaths = ["tests"]
